requirements.txt
streamlit
loguru
python-dotenv
openai
boto3
botocore
pillow

# python -m pip install --no-cache-dir --upgrade -r requirements.txt